chat/requirements.txt
flask==3.0.0
flask-socketio==5.3.5
redis==5.0.1
eventlet==0.33.3
python-socketio==5.10.0
python-dotenv==1.0.0
